+++ Website/data/locales/ja.json
@@ -42,7 +42,7 @@
 		"useSuggestedQuestion": "Do you want to use the suggested partner for this deck or select one yourself?",
 		"useSuggested": "Use Suggested Partner",
 		"selectManually": "Select One From Deck",
-		"waitingForOpponent": "Waiting for opponent to choose their partner...",
+		"waitingForOpponent": "Waiting for the opponent to choose their partner...",
 		"revealPartner": "Reveal Partner"
 	},
