+++ Website/index.html
@@ -477,6 +477,5 @@
 		<script src="scripts/ui.js"></script>
 		<script src="scripts/websocket.js"></script>
 		<script src="scripts/hotkeys.js"></script>
-		<script src="scripts/cardActions.js"></script>
 	</body>
 </html>
