+++ Website/modules/cardActions.js
@@ -1,6 +1,6 @@
-// This file holds info about actions that need to be accessible to specific cards.
+// This module exports a list of functions that need to be accessible as buttons on specific cards.
 
-let cardActions = {
+export let cardActions = {
 	"I00040": {
 		"roll": function() {
 			let result = Math.floor(Math.random() * 6) + 1;
