+++ Website/scripts/variables.js
@@ -1,27 +1,5 @@
 let game = null;
 let localPlayer = null;
-import("/modules/game.js").then(async gameModule => {
-	game = new gameModule.Game();
-	import("/modules/card.js").then(async cardModule => {
-		await fetch("https://crossuniverse.net/cardInfo", {
-			method: "POST",
-			body: JSON.stringify({
-				"cardTypes": ["token"],
-				"language": localStorage.getItem("language")
-			})
-		})
-		.then(response => response.json())
-		.then(response => {
-			response.forEach(card => {
-				card.imageSrc = getCardImageFromID(card.cardID);
-				game.cardData[card.cardID] = card;
-				cardAreas["tokens"].cards.push(new cardModule.Card(game, card.cardID));
-			});
-		});
-		localPlayer = game.players[1];
-	});
-});
-
 let gameState = null;
 
 // Areas that cards can go to. These handle card location changes and change the DOM accordingly
@@ -34,15 +12,6 @@ let roomCodeShown = false;
 let shiftHeld = false;
 let ctrlHeld = false;
 let altHeld = false;
-
-let officialDecks = [];
-fetch("data/deckList.json")
-.then(response => response.json())
-.then(decks => {
-	officialDecks = decks;
-});
-
-let currentDeckList = "default";
 
 let youAre = null; // Whether this client is player 0 or player 1. (Mainly for draft games and partner selection, as far as the board is concerned, the local player is always player 1.)
